requirements.txt
fastapi
uvicorn
sqlalchemy
anthropic
openai
python-dotenv
websockets
jinja2 